+++ Changelog.md
@@ -78,3 +78,5 @@
 0.7.8 balances problem
 0.7.9 balances problem
 0.8.0 balances problem
+0.8.1 balances problem
+0.8.2 balances problem